fix type issues
