+++ src/shared/options/augmentOptionsWithExcludes.ts
@@ -1,7 +1,7 @@
 import * as prompts from "@clack/prompts";
-import { z } from "zod";
 
 import { filterPromptCancel } from "../prompts.js";
+import { InputBase, Options } from "../types.js";
 
 const exclusionDescriptions = {
 	excludeCompliance: {
@@ -68,42 +68,6 @@ type ExclusionKey = keyof typeof exclusionDescriptions;
 
 const exclusionKeys = Object.keys(exclusionDescriptions) as ExclusionKey[];
 
-export const optionsSchema = z.object({
-	author: z.string().optional(),
-	base: z
-		.union([z.literal("everything"), z.literal("minimum"), z.literal("prompt")])
-		.optional(),
-	createRepository: z.boolean().optional(),
-	description: z.string(),
-	email: z.string().email().optional(),
-	excludeCompliance: z.boolean().optional(),
-	excludeContributors: z.boolean().optional(),
-	excludeLintJson: z.boolean().optional(),
-	excludeLintKnip: z.boolean().optional(),
-	excludeLintMd: z.boolean().optional(),
-	excludeLintPackageJson: z.boolean().optional(),
-	excludeLintPackages: z.boolean().optional(),
-	excludeLintPerfectionist: z.boolean().optional(),
-	excludeLintSpelling: z.boolean().optional(),
-	excludeLintYml: z.boolean().optional(),
-	excludeReleases: z.boolean().optional(),
-	excludeRenovate: z.boolean().optional(),
-	excludeTests: z.boolean().optional(),
-	funding: z.string().optional(),
-	owner: z.string(),
-	repository: z.string(),
-	skipGitHubApi: z.boolean(),
-	skipInstall: z.boolean().optional(),
-	skipRemoval: z.boolean().optional(),
-	skipRestore: z.boolean().optional(),
-	skipUninstall: z.boolean().optional(),
-	title: z.string(),
-});
-
-export type Options = z.infer<typeof optionsSchema>;
-
-type Base = NonNullable<Options["base"]>;
-
 export async function augmentOptionsWithExcludes(
 	options: Options,
 ): Promise<Options | undefined> {
@@ -119,7 +83,7 @@ export async function augmentOptionsWithExcludes(
 
 	const base =
 		options.base ??
-		filterPromptCancel<Base | symbol>(
+		filterPromptCancel<InputBase | symbol>(
 			await prompts.select({
 				message: `How much tooling would you like the template to set up for you?`,
 				options: [
